requirements.txt
# -----------------------------
# Core
# -----------------------------
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
requests==2.32.1
langchain==0.1.3

# -----------------------------
# Document processing
# -----------------------------
PyPDF2==3.0.1

# -----------------------------
# Embeddings
# -----------------------------
sentence-transformers==2.2.2
torch==2.2.0

# -----------------------------
# Vectorstore
# -----------------------------
faiss-cpu==1.11.0

# -----------------------------
# LLM generator
# -----------------------------
transformers==4.35.0

# -----------------------------
# Optional frontend
# -----------------------------
streamlit==1.29.0
